Make temperature fields float.

+++ pi_control/temperature/models.py
@@ -3,8 +3,8 @@ from django.db import models
 
 class Entry(models.Model):
     time = models.DateTimeField(auto_now_add=True, verbose_name='Measurment time.')
-    temperature = models.IntegerField(verbose_name='Temperature in Celsius.')
-    humidity = models.IntegerField(verbose_name='Humidity in %.')
+    temperature = models.FloatField(verbose_name='Temperature in Celsius.')
+    humidity = models.FloatField(verbose_name='Humidity in %.')
 
     class Meta:
         verbose_name = 'temperature entry'
